Update engine version
Some of my PRs that i needed here are merged.

Now Commando Actually has his C4 against vehicles.
And now there is a Sight crate bonus, current sight is also shown on the
UI. In original sight effect was a part of Range bonus, but i wanted to
seperate them.