Add Time Limit option.
When timer runs out one with the most points win. If there are equal
amount of players with same highest score, it shows Tie! until the
points change which causes the game to end.

Currently in a team game, team with most scoring player wins, rather
than team with most total score but i'll probably change that. This was
already annoying to code.