Implement a new spawning logic with Team Spawn lobby option.
Also sets most maps' player count to 50. New spawning code is set to not crash in that case.

There are 2 issues currently. Placement is always complately random. Instead of previous keeps players far from each other logic.

And the starting position options in the lobby is currently ignored. I may update the code to include it in the future.