Add More Anti-Air checbox
I've made a lot of units capable of attacking air for SS mod, but i think this is better to make it a checbox.

When enabled. Minigunner, Buggy, Humvee, Light Tank, Medium Tank and  Chinook gains anti-air capability.

Also fixed that L/MTNK didn't show an animation when hitting aircraft.